refactor(requests): share category cashflow direction resolution
Identical 24-line prepareForValidation() in Store/Update category
requests moved to ResolvesCategoryCashflowDirection trait.
