+++ app/Http/Requests/Settings/UpdateCategoryRequest.php
@@ -5,43 +5,21 @@ namespace App\Http\Requests\Settings;
 use App\Enums\CategoryCashflowDirection;
 use App\Enums\CategoryColor;
 use App\Enums\CategoryType;
+use App\Http\Requests\Concerns\ResolvesCategoryCashflowDirection;
 use Illuminate\Contracts\Validation\ValidationRule;
 use Illuminate\Foundation\Http\FormRequest;
 use Illuminate\Validation\Rule;
 
 class UpdateCategoryRequest extends FormRequest
 {
+    use ResolvesCategoryCashflowDirection;
+
     /**
      * Determine if the user is authorized to make this request.
      */
     public function authorize(): bool
     {
         return true;
-    }
-
-    protected function prepareForValidation(): void
-    {
-        $type = CategoryType::tryFrom((string) $this->input('type'));
-
-        if (in_array($type, [CategoryType::Savings, CategoryType::Investment], true)) {
-            $this->merge([
-                'cashflow_direction' => CategoryCashflowDirection::Outflow->value,
-            ]);
-
-            return;
-        }
-
-        if ($type !== CategoryType::Transfer) {
-            $this->merge([
-                'cashflow_direction' => CategoryCashflowDirection::Hidden->value,
-            ]);
-
-            return;
-        }
-
-        $this->merge([
-            'cashflow_direction' => $this->input('cashflow_direction', CategoryCashflowDirection::Hidden->value),
-        ]);
     }
 
     /**
